expanded info kept on hooks and generic functions
this had to be done in preperation for defining dispatchers for
different qualifiers.

a quick change towards the end of this commit has meant that the hooks
for a type specializer list has to be looked up at runtime. This has
to be the case for any method that has not been defined with
finalize-dispatch-method as if a hook is redefined to be qualified
differently, it's original method will not be updated to exclude the
hook.

There might be a work around to the above that define-hook will check
if the hook exists and redefine the old dispatch method but it's not
clear whether this will be worth at at this time. The main goal is to
just get to a stage where dispatch is different based on qualifier.

We probably should have done this more bottom up than top down than we
have currently.