This is a little messy details below
define-hook-function can only be used in the top-level for simple
reasons, the generic has to be defined before the hooks for it are.

The make load forms in this commit therefor are kind of unecessary and
need cleaning out, there's probably lots of load-forms here that are
not needed and we really need to devise a good method for figuring out
when to load things and specifically where to do that.

finilize-method-dispatch still needs testing in the new environment too.