new internal macros for finalize-dispatch-method
`dispatch` can be used to dispatch the hooks for the given generic and
type specializer list.

`%lay-method-base-for-dispatch` will be used to produce the shell of a
defmethod for a dispatcher so that both finalize-dispatch-method can
be used to change the way dispatch happens & so that there isn't
repeated code inside %define-dispatch-method.